feat: add routing and move App content to Home page
